+++ test/Test.cpp
@@ -94,6 +94,7 @@ void Test::run() {
     testFailedWrites();
     testTimestamp();
     testRetryOnFailedConnection();
+    testRetryOnFailedConnectionWithFlush();
     testBufferOverwriteBatchsize1();
     testBufferOverwriteBatchsize5();
     testServerTempDownBatchsize5();
@@ -498,6 +499,81 @@ void Test::testRetryOnFailedConnection() {
     String query = "select";
     FluxQueryResult q = clientOk.query(query);
     TEST_ASSERT(countLines(q) == 3);
+
+    TEST_END();
+    deleteAll(Test::apiUrl);
+}
+
+void Test::testRetryOnFailedConnectionWithFlush() {
+    TEST_INIT("testRetryOnFailedConnectionWithFlush");
+
+    InfluxDBClient clientOk(Test::apiUrl, Test::orgName, Test::bucketName, Test::token);
+    clientOk.setWriteOptions(WriteOptions().batchSize(2).bufferSize(2).retryInterval(4));
+    waitServer(Test::managementUrl, true);
+    TEST_ASSERT(clientOk.validateConnection());
+    Point *p = createPoint("test1");
+    TEST_ASSERT(clientOk.writePoint(*p));
+    delete p;
+    TEST_ASSERT(clientOk.flushBuffer());
+    TEST_ASSERT(clientOk.isBufferEmpty());
+
+    clientOk.setHTTPOptions(HTTPOptions().httpReadTimeout(500));
+
+    Serial.println("Stop server!");
+    waitServer(Test::managementUrl, false);
+    // test dropping batch on max retry count
+    TEST_ASSERT(!clientOk.validateConnection());
+    p = createPoint("test1");
+    TEST_ASSERT(clientOk.writePoint(*p));
+    delete p;
+
+    Serial.print(millis()/1000.0f,3);
+    Serial.println(" Write 1");
+
+    TEST_ASSERT(!clientOk.flushBuffer());
+    TEST_ASSERT(!clientOk.isBufferEmpty());
+    Serial.println(clientOk.getLastErrorMessage());
+    
+    Serial.print(millis()/1000.0f,3);
+    Serial.println(" Write 2");
+
+    TEST_ASSERT(!clientOk.flushBuffer());
+    TEST_ASSERT(!clientOk.isBufferEmpty());
+    Serial.println(clientOk.getLastErrorMessage());
+
+    Serial.print(millis()/1000.0f,3);
+    Serial.println(" Write 3");
+
+    TEST_ASSERT(!clientOk.flushBuffer());
+    TEST_ASSERT(!clientOk.isBufferEmpty());
+    Serial.println(clientOk.getLastErrorMessage());
+    
+   
+    Serial.println("Start server!");
+    waitServer(Test::managementUrl, true);
+    clientOk.setHTTPOptions(HTTPOptions().httpReadTimeout(5000));
+    TEST_ASSERT(clientOk.validateConnection());
+
+    Serial.print(millis()/1000.0f,3);
+    Serial.println(" Write 4");
+    p = createPoint("test1");
+    TEST_ASSERT(clientOk.writePoint(*p));
+    delete p;
+    TEST_ASSERT(clientOk.flushBuffer());
+    TEST_ASSERT(clientOk.isBufferEmpty());
+
+    Serial.print(millis()/1000.0f,3);
+    Serial.println(" Write 5");
+    p = createPoint("test1");
+    TEST_ASSERT(clientOk.writePoint(*p));
+    delete p;
+    TEST_ASSERT(clientOk.flushBuffer());
+    TEST_ASSERT(clientOk.isBufferEmpty());
+
+    String query = "select";
+    FluxQueryResult q = clientOk.query(query);
+    TEST_ASSERT(countLines(q) == 3);
+
 
     TEST_END();
     deleteAll(Test::apiUrl);