+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 3.5.1 [in progress]
 ### Fixes
 - [#114](https://github.com/tobiasschuerg/InfluxDB-Client-for-Arduino/pull/114) - Renamed `getRemaingRetryTime()`->`getRemainingRetryTime()`
+- [#115](https://github.com/tobiasschuerg/InfluxDB-Client-for-Arduino/pull/115) - Restored writing capability after a connection failure
 
 ## 3.5.0 [2020-10-30]
 ### Features